tests/wpt/web-platform-tests/origin-policy/features/comma-in-policy.https.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Commas in "features/policy" cause parse errors and thus no feature policy</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="helper.js"></script>

<script>
"use strict";
runFPTest({ camera: false, geolocation: false });
</script>
