third_party/blink/web_tests/virtual/origin-policy/external/wpt/origin-policy/features/comma-in-policy.https-expected.txt

This is a testharness.js-based test.
FAIL Commas in "features/policy" cause parse errors and thus no feature policy assert_equals: camera expected false but got true
Harness: the test ran to completion.
